refactor: remove intercom, add hubspot messages, move notifications to left, remove support button
