+++ app/app.js
@@ -1,8 +1,8 @@
 import Ember from 'ember';
 import Resolver from 'irene/resolver';
 import config from 'irene/config/environment';
-import installIntercom from 'irene/utils/install-intercom';
 import installPendo from 'irene/utils/install-pendo';
+import installHubspot from 'irene/utils/install-hubspot';
 import installInspectlet from 'irene/utils/install-inspectlet';
 import customerSuccessBox from 'irene/utils/customer-success-box';
 import loadInitializers from 'ember-load-initializers';
@@ -19,9 +19,9 @@ if (config.isAppknox) {
   config.product = ENUMS.PRODUCT.DEVKNOX;
 }
 
-installIntercom();
 customerSuccessBox();
 installPendo();
+installHubspot();
 installInspectlet();
 
 Ember.MODEL_FACTORY_INJECTIONS = true;
