+++ app/models/user.js
@@ -35,7 +35,6 @@ const User = DS.Model.extend({
   mfaMethod: DS.attr('number'),
   mfaSecret: DS.attr('string'),
   isTrial: DS.attr('boolean'),
-  intercomHash: DS.attr('string'),
 
   tProject: t("project"),
   tProjects: t("projects"),
