+++ app/routes/authenticated.js
@@ -43,20 +43,6 @@ const AuthenticatedRoute = Ember.Route.extend(AuthenticatedRouteMixin, {
     };
     triggerAnalytics('login', data);
     try {
-      window.Intercom("boot", {
-        app_id: ENV.intercomAppID,
-        name: user.get("username"),
-        email: user.get("email"),
-        alignment: 'left',
-        horizontal_padding: 20,
-        vertical_padding: 20,
-        custom_launcher_selector: '#intercom_support',
-        user_hash: user.get("intercomHash")
-      }
-      );
-      window.Intercom('trackEvent', 'logged-in');
-    } catch (error2) {}
-    try {
       const mixpanel = this.get("mixpanel");
       mixpanel.identify(user.get("id"));
       mixpanel.peopleSet({
