+++ app/templates/application.emblem
@@ -1,4 +1,4 @@
 = outlet
 = irene-onboard
-= notification-container position="bottom-right"
+= notification-container position="bottom-left"
 = namespace-modal
