+++ tests/unit/utils/install-hubspot-test.js
@@ -1,0 +1,10 @@
+import installHubspot from 'irene/utils/install-hubspot';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | install hubspot');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = installHubspot();
+  assert.ok(result);
+});
