--- tests/unit/utils/install-intercom-test.js
@@ -1,7 +1,0 @@
-import { module, test } from 'qunit';
-
-module('Unit | Utility | install intercom');
-
-test('it works', function(assert) {
-  assert.ok(true);
-});
